refactor: rename ListV2 folder
